Added seed output. Added complete prices (runsxtime matrix) and orders outputs.
Seed is now output in the overall output file at the end. This is to
verify determinism as well as to allow strange results to be reproduced.
Additionally complete prices/orders matrices are being output in
addition to other outputs in an effort to help reduce total outputs
size.
Added Seed class. This is a singleton which updates the seed and has an
access only seed value.
Added uses of the Seed class in OpenClDevice and OrderBookSim.
Added prices/orders outputs in Logger